requirements.txt
discord.py>=2.6.0
httpx
openai
pyyaml
google-genai
gemini-webapi
table2ascii
markdown2image
markdown
pandas
matplotlib
dataframe-image
